requirements.txt
streamlit
pandas
numpy
scikit-learn
plotly
mysql-connector-python
sqlalchemy

